requirements.txt
PyVirtualSerialPorts